chore(ci): Adds autoconfig and multikas tests
